Update new style.
