+++ README-CONTRIBUTE.md
@@ -1,0 +1,76 @@
+# 贡献
+
+# 感谢
+## 感谢您的宝贵建议
+由衷感谢他们对本文内容提出了自己的宝贵意见.  
+
+*排名不分前后. 简介内容与字数等不能代表任何主观含义. 若未及时收录不代表任何特殊意义.  *  
+
+| ID | 备注 |
+| -----   | -----   |
+| andylizi | 在本文撰写初期提出了诸多宝贵意见与指导. |  
+| Sotr | 针对本文多处措辞问题、叙述方式等提出诸多宝贵意见. |
+| Lss233 | 对本文不当内容提出宝贵意见. |
+| frea | 指出了本文中一处较为重要的错别字. |
+| hcrgm | 为本文提出诸多指导意见. |
+| Henry | 为本文翻译问题提出了宝贵意见. |
+| 602723113 | 为本文粒子特效部分提供了宝贵的帮助. |
+| USED TO IT | 为教程指正了诸多严重的低级错误. |
+
+# 贡献本项目
+
+我们的内容尚未完成, 十分欢迎您为本项目贡献内容!
+
+## 对贡献人的要求
+
+我们希望您能满足(非硬性要求):
+
+1. 能够独立的编写出BukkitAPI插件, 且发表过一些插件.  
+2. 参与编写过与贡献内容相关的代码(最好对外发表、开源过).  
+3. 具有编写相关经验.  
+
+并满足如下条件(硬性要求):
+
+1. 对Java命名规则、团队开发常识等有一定认识.  
+2. 有Minecraft经验.
+3. 对Java面向对象具有一定的了解, 较为掌握Java语言.  
+4. 具有编写BukkitAPI插件的能力.  
+
+## 参与
+
+欢迎您`Fork`我们的GitHub项目, 创建`Pull Request`.  
+
+在贡献之前, 请您悉知:
+
+**若是章节结构安排问题.**
+
+```
+_sidebar.md    README.md   README-*.md
+```
+
+我们不允许您直接修改上述文件, 我们也不允许您为网站样式等问题直接作出贡献.  
+如果希望向我们提出更好的意见使内容编排更合理, 或为我们贡献更合理的网站样式, 请发issue.  
+
+目前我们正在使用Docsify.  
+我们不接受任何更换Docsify为GitBook、Typecho、Wordpress等不合理请求.  
+
+**若是修改现存章节内容 & 贡献尚未补充内容.**
+
+您可以按要求进行修改或补充, 完毕后请创建`Pull Request`.
+
+# 准则
+
+## 底线
+
+1. 代码应当先测试, 后使用, 代码格式标准以Eclipse或IDEA默认设置下的格式化结果为标准, 不允许出现Tab符.
+2. 应当注意代码是否适用于所有的BukkitAPI版本, 应当考虑在不同版本下的不同解决方式.  
+3. 应当注意文章内容是否合理.  
+4. 内容应避免敏感内容(例如违法内容、政**治内容、广告内容、恶意辱骂、不友好不实不适内容等).
+
+## 标准
+
+代码与内容应以如下标准编写:
+
+1. Java8环境, 可以使用Lambda.  
+2. 格式上应该沿袭之前的内容.
+3. 逗号、句号、分号标点应使用英文标点.
