+++ README-DONATE.md
@@ -1,0 +1,19 @@
+# 赞助
+
+# 赞助渠道
+
+爱发电: https://afdian.net/@tdiant
+
+# 赞助者名单
+
+本名单不会及时更新, 但一定会更新. 排序以时间顺序为基准.  
+若有遗漏、如需更正请联系我们.  
+*较为久远的内容已经难以查询, 若有需要请联系我们.*  
+*货币单位: 人民币(扣除手续费前)*
+
+ID|渠道|联系方式|时间|金额|备注
+-|-|-|-|-|-
+豆沙半糖|爱发电|.|2020年2月19日|5.00|.  
+
+
+
