+++ README.md
@@ -1,43 +1,41 @@
-# 欢迎！
+# 欢迎
+
+<div>
+    <a href="/#/_sidebar.md"><div style="width:25%;height:50px;background:#3388ff;border-style:none solid none none;border-color:white;border-width:2px;float:left;">
+        <span style="font-size:25px;margin-top:10px;position:absolute;text-align:center;color:white;float:left;margin-left:10px;">目录</span>
+        <span style="font-size:25px;margin-top:10px;text-align:center;color:white;float:right;margin-right:10px;">   <span class="fas fa-align-justify"></span></span>
+    </div></a>
+    <a href="https://github.com/tdiant/BukkitDevelopmentNote"><div style="width:25%;height:50px;background:#3388ff;border-style:none solid none none;border-color:white;border-width:2px;float:left">
+        <span style="font-size:25px;margin-top:10px;position:absolute;text-align:center;color:white;float:left;margin-left:10px;">Git</span>
+        <span style="font-size:25px;margin-top:10px;text-align:center;color:white;float:right;margin-right:10px;">   <span class="fas fa-code-branch"></span></span>
+    </div></a>
+    <a href="/#/README-CONTRIBUTE.md"><div style="width:25%;height:50px;background:#3388ff;border-style:none solid none none;border-color:white;border-width:2px;float:left">
+        <span style="font-size:25px;margin-top:10px;position:absolute;text-align:center;color:white;float:left;margin-left:10px;">贡献</span>
+        <span style="font-size:25px;margin-top:10px;text-align:center;color:white;float:right;margin-right:10px;">   <span class="fas fa-pen-nib"></span></span>
+    </div></a>
+    <a href="/#/README-DONATE.md"><div style="width:25%;height:50px;background:#3388ff;float:left">
+        <span style="font-size:25px;margin-top:10px;position:absolute;text-align:center;color:white;float:left;margin-left:10px;">赞助</span>
+        <span style="font-size:25px;margin-top:10px;text-align:center;color:white;float:right;margin-right:10px;">   <span class="fas fa-trophy"></span></span>
+    </div></a>
+</div>
+
 感谢您的访问.  
-本站旨在为BukkitAPI初学者提供简练, 有效的入门资料.  
+本站旨在为BukkitAPI开发者提供清晰的入门资料.  
 
 **阅读本文需要Java基础.**
-
-# 聊天
-来这里一起HAPPY!(需要活跃气氛的, 最好是那种特别能聊天的  
-Q群: 861294270  
-~~我常年住校! 不一定有空回答你问题! emmmmmmm~~
 
 # 版权声明
 本站内容作者为 tdiant **和以下所感谢的这些朋友**. 部分文献资料若涉及转载或引用, 将会在相关内容处标记.  
 
 文档中代码均以Apache 2.0开源, 包括正文部分插入代码.  
-<a rel="license" href="http://creativecommons.org/licenses/by-nc-nd/3.0/cn/"><img alt="知识共享许可协议" style="border-width:0" src="https://i.creativecommons.org/l/by-nc-nd/3.0/cn/88x31.png" /></a><br />本作品采用<a rel="license" href="http://creativecommons.org/licenses/by-nc-nd/3.0/cn/">知识共享署名-非商业性使用-禁止演绎 3.0 中国大陆许可协议</a>进行许可。
+本作品采用<a rel="license" href="http://creativecommons.org/licenses/by-nc-nd/3.0/cn/">知识共享署名-非商业性使用-禁止演绎 3.0 中国大陆许可协议</a>进行许可。
 
-若需转载任何内容，请详细注明出处.   
+若需转载任何内容, 请详细注明出处, 仅供学习交流使用, 谢绝商业化使用.  
 
 本站图片资源部分来自互联网，仅学习交流使用。部分图片由**tdiant**绘制。  
-若图片需要转载使用，仅供学习交流使用。
 
-# 感谢
-## 感谢您的宝贵建议
-由衷感谢他们对本文内容提出了自己的宝贵意见.  
-
-*排名不分前后. 简介内容与字数等不能代表任何主观含义. 若未及时收录不代表任何特殊意义.  *  
-
-| ID | 备注 |
-| -----   | -----   |
-| andylizi | 在本文撰写初期提出了诸多宝贵意见与指导. |  
-| Sotr | 针对本文多处措辞问题、叙述方式等提出诸多宝贵意见. |
-| Lss233 | 对本文不当内容提出宝贵意见. |
-| frea | 指出了本文中一处较为重要的错别字. |
-| hcrgm | 为本文提出诸多指导意见. |
-| Henry | 为本文翻译问题提出了宝贵意见. |
-| 602723113 | 为本文粒子特效部分提供了宝贵的帮助. |
-| USED TO IT | 为教程指正了诸多严重的低级错误. |
-
-## 感谢您的打赏
-由衷感谢他们的打赏, 这是对我最好的鼓励!  
-如果你希望打赏, 请访问: https://afdian.net/@tdiant
+# 聊天
+来这里一起HAPPY!(需要活跃气氛的, 最好是那种特别能聊天的  
+Q群: 861294270  
+~~我常年住校! 不一定有空回答你问题! emmmmmmm~~
