+++ _sidebar.md
@@ -1,35 +1,34 @@
 * [欢迎](README.md)
 * BukkitAPI入门
-  - [开始](brm-0-1.md)
-  - 第一单元 基础
-    - [第一节 最简单的插件](brm-1-1.md)
-    - [第二节 事件](brm-1-2.md)
-    - [第三节 对象化的MC](brm-1-3.md)
-    - [第四节 配置API](brm-1-4.md)
-    - [第五节 命令执行器](brm-1-5.md)
-    - [第六节 插件制作实例](brm-1-6.md)
-  - 第二单元 深入Bukkit
-    - [第一节 插件的加载机制](brm-2-1.md)
-    - [第二节 自定义事件](brm-2-2.md)
-    - [第三节 深入plugin.yml](brm-2-3.md)
-    - [第四节 配置API的序列化和遍历](brm-2-4.md)
-    - [第五节 Bukkit 的多线程多任务框架](brm-2-5.md)
-    - [第六节 自定义合成表](brm-2-6.md)
-    - [第七节 粒子效果 及 音效播放](brm-2-7.md)
-    - [第八节 世界生成器](brm-2-8.md)
-  - 第三单元 常见的其他第三方API
-    - 第一节 深入依赖(未完成)
-    - [第二节 经济 Vault](brm-3-2.md)
+  - [开始](brm/brm-0-1.md)
+  - 第一单元 基础内容
+    - [第一节 最简单的插件](brm/brm-1-1.md)
+    - [第二节 事件](brm/brm-1-2.md)
+    - [第三节 对象化的MC](brm/brm-1-3.md)
+    - [第四节 配置API](brm/brm-1-4.md)
+    - [第五节 命令执行器](brm/brm-1-5.md)
+    - [第六节 简单的插件制作实例](brm/brm-1-6.md)
+  - 第二单元 进阶功能
+    - 第一节 Bukkit类与箱子GUI的实现(未完成)
+    - [第二节 自定义事件](brm/brm-2-2.md)
+    - [第三节 深入plugin.yml](brm/brm-2-3.md)
+    - [第四节 配置API的序列化和遍历](brm/brm-2-4.md)
+    - [第五节 Bukkit 的多线程多任务框架](brm/brm-2-5.md)
+    - [第六节 自定义合成表](brm/brm-2-6.md)
+    - [第七节 粒子效果 及 音效播放](brm/brm-2-7.md)
+    - [第八节 世界生成器](brm/brm-2-8.md)
+  - 第三单元 NMS与OBC
+    - [第一节 认识NMS与OBC](brm/brm-4-1.md)
+    - 第二节 练习反射(未完成)
+	  - 第三节 发包(未完成)
+	  - 第四节 操作NBT数据(未完成)
+  - 第四单元 常见的其他第三方API
+    - [第一节 深入依赖](brm/brm-3-1.md)
+    - [第二节 经济 Vault](brm/brm-3-2.md)
     - 第三节 跨服 Bungeecord(未完成)
     - 第四节 占位 PlaceholderAPI(未完成)
-* 插件开发进阶(未完成)
-  - 第一单元 NMS与OBC
-    - [第一节 NMS与OBC的基本概念](bsd-1-1.md)
-	- 第二节 发包
-	  - PART1 利用ProtocolLib发包(未完成)
-	  - PART2 不利用ProtocolLib发包(未完成)
-	- 第三节 操作NBT数据(未完成)
+* 插件开发实战
+  - [最简单的登录插件 StupidLogin](brm/brm-1-6.md)
 * 神奇的异次元
   - [Conversation API](rua-1.md)
-  
-	
+  - 插件使用统计的接入(未完成)
