--- brm-2-1.md
@@ -1,30 +1,0 @@
-# 插件的加载机制
-
-插件的加载顺序会因为多种原因而改变.  
-
-# 依赖
-如果我们的插件需要调用其他插件的API, 例如获取经济插件中玩家金币数据, 那么我们的插件就会依赖(depend)那个插件.  
-我们可以在`plugin.yml`中设置依赖, Bukkit会把 我们依赖的插件 放在 我们的插件 之前被加载.  
-
-有两种依赖方式:  
-1. depend: 我需要依赖其他插件, 并且我必须依赖它.  
-2. softdepend: 我需要依赖其他插件, 但这是可选的, 不依赖该插件仍能正常工作.  
-
-*在后续将更进一步地研究依赖.*
-
-# onLoad、onEnable和onDisable  
-在HelloWorld插件的实验中可得, `onEnable`会在服务器启动时被加载, `onDisable` 会在服务器关闭时被加载.  
-事实上, 还有第三个方法, `onLoad`.  
-
-这三个方法的关系如下:  
-**对于服务端启动**:  
-1. 服务端启动, 服务端体系加载完毕后, 加载所有插件jar文件, 按照顺序逐个调用每个插件的`onLoad`方法. 
-2. 所有插件`onLoad`方法调用完毕. 服务端在一些处理过后, 按照顺序逐个调用每个插件的`onEnable`方法.    
-
-**对于服务端关闭(stop指令)**:  
-服务端关闭时, 逐一调用每个插件的`onDisable`方法.
-
-**对于reload指令**:  
-1. reload指令触发后, 服务端在一些处理后, 按照顺序逐个调用每个插件的`onDisable`方法.    
-2. 所有插件`onDisable`方法调用完毕, 服务端按照顺序逐个调用每个插件的`onLoad`方法.  
-3. 所有插件`onLoad`方法调用完毕, 服务端按照顺序逐个调用每个插件的`onEnable`方法.
