+++ index.html
@@ -6,7 +6,16 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
   <meta name="description" content="Description">
   <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-  <link rel="stylesheet" href="//unpkg.com/docsify/lib/themes/vue.css">
+  <link rel="stylesheet" href="//unpkg.com/docsify/lib/themes/buble.css">
+  
+  <link href="https://cdn.bootcss.com/font-awesome/5.12.1/css/all.min.css" rel="stylesheet">
+
+  <!--<style>
+    .sidebar{
+      width:auto;
+      left:-22rem;
+    }
+  </style>-->
 </head>
 <body>
     <div id="app">Loading...... Please wait.</div>
@@ -17,24 +26,30 @@
       loadSidebar: true,
       subMaxLevel: 3,
 
-      search: {
-        maxAge: 86400000,               // 过期时间，单位毫秒，默认一天
-        paths: [],                      // or 'auto'，匹配文件路径
-        placeholder: {
-           '/zh-cn/': '搜索',
-           '/': 'Type to search'
-        },
-        noData: {
-           '/zh-cn/': '找不到结果',
-           '/': 'No Results'
-        },
-        depth: 4,                       // 搜索标题的最大程级, 1 - 6
-      }
+search: {
+  maxAge: 86400000,               // 过期时间，单位毫秒，默认一天
+  paths: [],                      // or 'auto'，匹配文件路径
+  placeholder: {
+     '/zh-cn/': '搜索',
+     '/': 'Type to search'
+  },
+  noData: {
+     '/zh-cn/': '找不到结果',
+     '/': 'No Results'
+  },
+  depth: 4,                       // 搜索标题的最大程级, 1 - 6
+}
     }
   </script>
   <script src="//unpkg.com/docsify/lib/docsify.min.js"></script>
   <script src="//unpkg.com/docsify/lib/plugins/search.js"></script>
   <script src="//unpkg.com/prismjs/components/prism-java.min.js"></script>
   <script src="//unpkg.com/prismjs/components/prism-yaml.min.js"></script>
+
+  <script>
+    (document.getElementsByClassName("sidebar-toggle-button")[0].onclick=function(){
+      console.log("aa");
+    });
+  </script>
 </body>
 </html>
