--- source/master_1-6/StupidLogin/src/plugin.yml
@@ -1,13 +1,0 @@
-name: StupidLogin
-main: tdiant.bukkit.stupidlogin.StupidLogin
-version: 1
-author: tdiant
-
-commands:
-  login:
-    usage: "/login YOUR_PASSWORD"
-    description: "Login your account."
-  register:
-    usage: "/register YOUR_PASSWORD"
-    description: "Register your account."
-
